Quick fix for kerning problem
The problem was caused by int64 logic being used. Emscripten doesn't
really like that.

+++ src/hb-ot-layout-common-private.hh
@@ -795,7 +795,7 @@ struct Device
 
     if (!pixels) return 0;
 
-    return pixels * (int64_t) scale / ppem;
+    return (int) (pixels * (double) scale / ppem);
   }
